requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-classic
faiss-cpu
pypdf
python-dotenv
streamlit
sentence-transformers
motifer==1.1.3
